iterate makePlots

+++ makePlots/go_plots.sh
@@ -13,7 +13,7 @@ export SCRAM_ARCH=slc5_amd64_gcc462
 eval `scramv1 runtime -sh`
 
 FILE_TO_RUN=$1
-cat makePlots_mc_template.C | sed s:FILE_TO_RUN:$FILE_TO_RUN: > makePlots.C
+cat makePlots_template.C | sed s:FILE_TO_RUN:$FILE_TO_RUN: > makePlots.C
 root -b -q -l makePlots.C | sed '/.root does not exist/d'
 rm makePlots.C
 
@@ -23,29 +23,30 @@ rm signal_contamination_stop.root
 rm contamination_stop.root
 rm plots_*.root
 
-#for x in nojet j b jj bj bb jjj bjj bbj 4j 4j1b 4j2b 5j 5j1b 5j2b 6j 6j1b 6j2b l lj lb ljj lbj lbb ljjj lbjj lbbj ljjjj lbjjj lbbjj ll llj llb lljj llbj llbb
-for x in nojet j b bj muJets eleJets hadronic
-do
-
-	cp template_errorTable_mc.tex errorTable_$x.tex
-
+if [ 1 -eq 0 ]; then
+    for x in ele
+    do
+	
+	cp template_errorTable.tex errorTable_$x.tex
+	
 	while read line
 	do
-		code=`echo $line | cut -d : -f 1`
-		value=`echo $line | cut -d : -f 2`
-
-		sed -i "s/${code}/${value}/g" errorTable_$x.tex
+	    code=`echo $line | cut -d : -f 1`
+	    value=`echo $line | cut -d : -f 2`
+	    
+	    sed -i "s/${code}/${value}/g" errorTable_$x.tex
 	done < errorTable_$x.temp
-
+	
 	rm errorTable_$x.temp
-
+	
 	latex errorTable_$x.tex
 	dvips -Ppdf -t landscape errorTable_$x.dvi
 	ps2pdf errorTable_$x.ps
-
+	
 	rm errorTable_$x.log errorTable_$x.dvi errorTable_$x.aux errorTable_$x.ps
-
-done
+	
+    done
+fi
 
 hadd contamination_stop.root signal_*.root
 rm signal_*.root
